feat: move qr code glue in MSALoginDialog
