+++ .gitmodules
@@ -19,6 +19,6 @@
 [submodule "flatpak/shared-modules"]
 	path = flatpak/shared-modules
 	url = https://github.com/flathub/shared-modules.git
-[submodule "libraries/qt-qrcodegenerator/QR-Code-generator"]
-	path = libraries/qt-qrcodegenerator/QR-Code-generator
+[submodule "libraries/qrcodegenerator"]
+	path = libraries/qrcodegenerator
 	url = https://github.com/nayuki/QR-Code-generator
