+++ flake.nix
@@ -16,7 +16,7 @@
       flake = false;
     };
 
-    qt-qrcodegenerator = {
+    qrcodegenerator = {
       url = "github:nayuki/QR-Code-generator";
       flake = false;
     };
@@ -27,7 +27,7 @@
       self,
       nixpkgs,
       libnbtplusplus,
-      qt-qrcodegenerator,
+      qrcodegenerator,
     }:
 
     let
@@ -175,7 +175,7 @@
         prismlauncher-unwrapped = prev.callPackage ./nix/unwrapped.nix {
           inherit
             libnbtplusplus
-            qt-qrcodegenerator
+            qrcodegenerator
             self
             ;
         };
